refactor: use lucide icons directly
- Replace duplicated UI icon components with lucide imports

- Share parent-path BackButton from components/ui
