+++ src/app/posts/[slug]/_components/footer.tsx
@@ -1,11 +1,11 @@
 'use client';
 
+import { BackButton } from '@/components/ui/backButton';
 import { ContentFooter } from '@/components/ui/contentFooter';
 import { ROUTES } from '@/constants/menu.constants';
 import { METADATA } from '@/constants/metadata.constants';
 
 import type { Post } from '@/types/post.types';
-import { BackButton } from './back-button';
 
 export const Footer = ({ slug }: Post) => {
   const url = `${METADATA.SITE.URL}${ROUTES.POSTS}/${slug}`;
