+++ src/app/projects/[slug]/_components/footer.tsx
@@ -1,10 +1,10 @@
 'use client';
 
+import { BackButton } from '@/components/ui/backButton';
 import { ContentFooter } from '@/components/ui/contentFooter';
 import { ROUTES } from '@/constants/menu.constants';
 import { METADATA } from '@/constants/metadata.constants';
 import type { Project } from '@/types/project.types';
-import { BackButton } from './back-button';
 
 export const Footer = ({ slug }: Project) => {
   const url = `${METADATA.SITE.URL}${ROUTES.PROJECTS}/${slug}`;
